add test_plan_statistics to test_camp and test_plan
